Move uncommon methods from component to service
