+++ backend/src/main/java/com/sprawler/external/myinfo/MyInfoController.java
@@ -1,5 +1,7 @@
 package com.sprawler.external.myinfo;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.nimbusds.jose.JOSEException;
 import com.sprawler.external.myinfo.dto.request.TokenRequestDTO;
 import com.sprawler.external.myinfo.entity.person.decrypted.DecryptedPersonInfo;
 import com.sprawler.external.myinfo.entity.token.TokenApiResponse;
@@ -9,7 +11,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.*;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.Map;
+import java.net.MalformedURLException;
+import java.security.NoSuchAlgorithmException;
+import java.security.spec.InvalidKeySpecException;
+import java.text.ParseException;
 
 @RestController("myInfoController")
 @RequestMapping("/myinfo")
@@ -28,7 +33,7 @@ public class MyInfoController {
     }
 
     @GetMapping("/authorize")
-    public ResponseEntity<String> makeAuthorizeCall(@RequestParam("verifier") String verifier) {
+    public ResponseEntity<String> makeAuthorizeCall(@RequestParam("verifier") String verifier) throws NoSuchAlgorithmException {
         LOGGER.info("Making api call to authorize data");
 
         String redirectUrl = myInfoService.createAuthRedirectUrl(verifier);
@@ -39,14 +44,14 @@ public class MyInfoController {
 
     @PostMapping(path = "/token", consumes = MediaType.APPLICATION_JSON_VALUE)
     public TokenApiResponse obtainMyInfoAccessToken(
-            @RequestBody TokenRequestDTO tokenRequestDTO) {
+            @RequestBody TokenRequestDTO tokenRequestDTO) throws NoSuchAlgorithmException, InvalidKeySpecException, JOSEException {
         return myInfoService.retrieveAccessToken(tokenRequestDTO);
     }
 
     @GetMapping("/person")
     public DecryptedPersonInfo getPersonData(
             @RequestParam("access_token") String accessToken,
-            @RequestParam("dpop_string") String dpopString) {
+            @RequestParam("dpop_string") String dpopString) throws NoSuchAlgorithmException, InvalidKeySpecException, JsonProcessingException, MalformedURLException, ParseException, JOSEException {
 
         return myInfoService.getPersonDataset(accessToken, dpopString);
     }
