+++ backend/src/main/java/com/sprawler/external/myinfo/MyInfoService.java
@@ -3,13 +3,24 @@ package com.sprawler.external.myinfo;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.nimbusds.jose.*;
+import com.nimbusds.jose.crypto.ECDHDecrypter;
+import com.nimbusds.jose.crypto.ECDSASigner;
+import com.nimbusds.jose.jwk.Curve;
 import com.nimbusds.jose.jwk.ECKey;
+import com.nimbusds.jose.jwk.KeyUse;
+import com.nimbusds.jose.jwk.gen.ECKeyGenerator;
 import com.nimbusds.jose.util.Base64URL;
+import com.nimbusds.jwt.EncryptedJWT;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
+import com.nimbusds.oauth2.sdk.id.JWTID;
 import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.oauth2.sdk.token.DPoPAccessToken;
 import com.sprawler.external.myinfo.dto.request.TokenRequestDTO;
 import com.sprawler.external.myinfo.entity.person.decrypted.DecryptedPersonInfo;
 import com.sprawler.external.myinfo.entity.token.TokenApiResponse;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,16 +35,14 @@ import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 
-import java.net.URLEncoder;
+import java.net.MalformedURLException;
 import java.nio.charset.StandardCharsets;
-import java.security.KeyFactory;
 import java.security.MessageDigest;
-import java.security.PrivateKey;
+import java.security.NoSuchAlgorithmException;
 import java.security.interfaces.ECPrivateKey;
-import java.security.spec.PKCS8EncodedKeySpec;
-import java.util.Base64;
-import java.util.Collections;
-import java.util.Map;
+import java.security.spec.InvalidKeySpecException;
+import java.text.ParseException;
+import java.util.*;
 
 @Service
 public class MyInfoService {
@@ -70,9 +79,9 @@ public class MyInfoService {
                 DecryptedPersonInfo.class).getBody();
     }
 
-    public String createAuthRedirectUrl(String verifier) {
+    public String createAuthRedirectUrl(String verifier) throws NoSuchAlgorithmException {
 
-        String codeChallenge = myInfoSecurity.createCodeChallenge(verifier);
+        String codeChallenge = createCodeChallenge(verifier);
 
         return UriComponentsBuilder
                 .fromUriString(authApiUrl)
@@ -84,15 +93,23 @@ public class MyInfoService {
                 .queryParam("code_challenge_method", codeChallengeMethod)
                 .queryParam("purpose_id", purposeId)
                 .toUriString();
-
     }
 
-    public TokenApiResponse retrieveAccessToken(TokenRequestDTO tokenRequestDTO) {
+    private String createCodeChallenge(String codeVerifier) throws NoSuchAlgorithmException {
+        byte[] bytes = codeVerifier.getBytes(StandardCharsets.US_ASCII);
+        MessageDigest sha256messageDigester = MessageDigest.getInstance("SHA-256");
+        sha256messageDigester.update(bytes, 0, bytes.length);
+        byte[] mdDigest = sha256messageDigester.digest();
+
+        return Base64.getUrlEncoder().withoutPadding().encodeToString(mdDigest);
+    }
+
+    public TokenApiResponse retrieveAccessToken(TokenRequestDTO tokenRequestDTO) throws JOSEException, NoSuchAlgorithmException, InvalidKeySpecException {
         LOGGER.info("Making api call to obtain token");
 
         LOGGER.info("Generating DPoP");
 
-        ECKey sessionPopKeyPair = myInfoSecurity.generateEphemeralKeys();
+        ECKey sessionPopKeyPair = generateEphemeralKeys();
         String dpopString = myInfoSecurity.generateDPoP(
                 tokenApiUrl,
                 "POST",
@@ -100,39 +117,67 @@ public class MyInfoService {
                 null,
                 null);
 
-        Base64URL jktThumbprint = null;
-        ECPrivateKey privateSigningKey = null;
-        try {
-            jktThumbprint = sessionPopKeyPair.toPublicJWK().computeThumbprint("SHA-256");
+        Base64URL jktThumbprint = sessionPopKeyPair.toPublicJWK().computeThumbprint("SHA-256");
+        String privateKeyStr = "MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgZw4GTT/we1cBHxeRKKcalUTgZOZs9CeOSEY74D5Cs9qgCgYIKoZIzj0DAQehRANCAAQFdRarRnZFEWquVZtbaa4jJs2eP9gHF+U1BMQ6D5CYJL1zuBuMg6NIjlSQM5hFsiGcF8J0pxsTyK2Xu0l9Dyru";
 
-            String privateKeyStr = "MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgZw4GTT/we1cBHxeRKKcalUTgZOZs9CeOSEY74D5Cs9qgCgYIKoZIzj0DAQehRANCAAQFdRarRnZFEWquVZtbaa4jJs2eP9gHF+U1BMQ6D5CYJL1zuBuMg6NIjlSQM5hFsiGcF8J0pxsTyK2Xu0l9Dyru";
-
-            byte[] keyBytes = Base64.getDecoder().decode(privateKeyStr);
-
-            // Step 2: Convert to PKCS8EncodedKeySpec
-            PKCS8EncodedKeySpec  keySpec = new PKCS8EncodedKeySpec(keyBytes);
-
-            // Step 3: Generate ECPrivateKey from KeyFactory
-            KeyFactory keyFactory = KeyFactory.getInstance("EC");
-            PrivateKey privateKey = keyFactory.generatePrivate(keySpec);
-
-            // Return as ECPrivateKey
-            privateSigningKey = (ECPrivateKey) privateKey;
-        } catch (Exception e) {
-            LOGGER.error(e);
-        }
+        ECPrivateKey privateSigningKey = myInfoSecurity.getEcPrivateKey(privateKeyStr);
 
         assert jktThumbprint != null;
 
-        LOGGER.info("Generating Client Assertion");
-        String clientAssertion = myInfoSecurity
-                .generateClientAssertion(
+        String clientAssertion = generateClientAssertion(
                         tokenApiUrl,
                         clientId,
                         jktThumbprint,
                         keyId,
                         privateSigningKey
                 );
+
+        HttpEntity<MultiValueMap<String, String>> requestEntity = generateTokenRequestEntity(tokenRequestDTO, dpopString, clientAssertion);
+
+        TokenApiResponse tokenApiResponseObject = myInfoTemplate
+                .postForEntity(tokenApiUrl, requestEntity, TokenApiResponse.class)
+                .getBody();
+
+        assert tokenApiResponseObject != null;
+
+        LOGGER.info("Decoding provided auth token to extract details");
+        return craftTokenApiResponse(tokenApiResponseObject, sessionPopKeyPair);
+    }
+
+    private ECKey generateEphemeralKeys() throws JOSEException {
+        return new ECKeyGenerator(Curve.P_256)
+                .keyID(RandomStringUtils.secure().next(40))
+                .algorithm(JWSAlgorithm.ES256)
+                .keyUse(KeyUse.SIGNATURE)
+                .generate();
+    }
+
+    private String generateClientAssertion(final String url, final String clientId, Base64URL jktThumbprint, String keyId, ECPrivateKey privateSigningKey) throws JOSEException {
+        Map<String, Object> cnf = new HashMap<>();
+        cnf.put("jkt", jktThumbprint.toString());
+
+        final JWSSigner signer = new ECDSASigner(privateSigningKey, Curve.P_256);
+        final SignedJWT signedJWT = new SignedJWT(
+                new JWSHeader.Builder(JWSAlgorithm.ES256)
+                        .keyID(keyId)
+                        .type(JOSEObjectType.JWT).build(),
+                new JWTClaimsSet.Builder()
+                        .issuer(clientId)
+                        .subject(clientId)
+                        .audience(url)
+                        .issueTime(new Date())
+                        .expirationTime(new Date(System.currentTimeMillis() + 300000L))
+                        .jwtID(new JWTID().getValue())
+                        .claim("cnf", cnf).build());
+
+        signedJWT.sign(signer);
+
+        return signedJWT.serialize();
+    }
+
+
+    private HttpEntity<MultiValueMap<String, String>> generateTokenRequestEntity(TokenRequestDTO tokenRequestDTO, String dpopString, String clientAssertion) {
+        LOGGER.info("Generating Client Assertion");
 
         // Set up headers
         HttpHeaders headers = new HttpHeaders();
@@ -153,14 +198,10 @@ public class MyInfoService {
         LOGGER.info(formParams);
         LOGGER.info(headers);
 
-        HttpEntity<MultiValueMap<String, String>> requestEntity = new HttpEntity<>(formParams, headers);
+        return new HttpEntity<>(formParams, headers);
+    }
 
-        TokenApiResponse tokenApiResponseObject = myInfoTemplate
-                .postForEntity(tokenApiUrl, requestEntity, TokenApiResponse.class)
-                .getBody();
-
-
-        LOGGER.info("Decoding provided auth token to extract details");
+    private TokenApiResponse craftTokenApiResponse(TokenApiResponse tokenApiResponseObject, ECKey sessionPopKeyPair) {
         String jwksUrl = "https://test.authorise.singpass.gov.sg/.well-known/keys.json";
         DecodedJWT tokenJWT = myInfoSecurity
                 .decodeJwtToken(tokenApiResponseObject.getAccess_token(), jwksUrl);
@@ -179,7 +220,7 @@ public class MyInfoService {
         return tokenApiResponseObject;
     }
 
-    public DecryptedPersonInfo getPersonDataset(String accessToken, String dpopString) {
+    public DecryptedPersonInfo getPersonDataset(String accessToken, String dpopString) throws NoSuchAlgorithmException, InvalidKeySpecException, JsonProcessingException, MalformedURLException, ParseException, JOSEException {
         LOGGER.info("Running api to retrieve person data");
 
         LOGGER.info("Decoding provided auth token");
@@ -196,18 +237,13 @@ public class MyInfoService {
         HttpEntity<MultiValueMap<String, String>> requestEntity = new HttpEntity<>(headers);
 
         // Build URI with path parameter
-
+        String personUri = UriComponentsBuilder
+                .fromUriString(personApiUrl)
+                .pathSegment(tokenJWT.getSubject())
+                .toUriString() + "?scope=" + scope;
 
         String encryptedResponse = "";
         try {
-            String personUri = UriComponentsBuilder
-                    .fromUriString(personApiUrl)
-                    .pathSegment(tokenJWT.getSubject())
-                    .toUriString() + "?scope=" + scope;
-
-            LOGGER.info(personUri);
-            LOGGER.info(headers.get("Authorization"));
-            LOGGER.info(headers.get("DPoP"));
 
             encryptedResponse = myInfoTemplate.exchange(
                     personUri,
@@ -220,52 +256,44 @@ public class MyInfoService {
         }
 
         LOGGER.info("String obtained. Attempting to decrypt");
-        LOGGER.info(encryptedResponse);
         String encryptionPrivateKeyString = "MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgOvpIsduu5J9C4s7G9zqcAvfD78OD0LemB7Voa5nyufagCgYIKoZIzj0DAQehRANCAAQ18jEm50E7X9cPgOeGRUo+ice7HQsbXY7cP5nY1+y4QdhxunsiXBBAVWVKdWJReey1jJQCJskrjecEHHhK5qPf";
 
         LOGGER.info("Loading Private Key");
-        ECPrivateKey encryptionPrivateKey = null;
-        try {
-
-
-            byte[] keyBytes = Base64.getDecoder().decode(encryptionPrivateKeyString);
-
-            // Step 2: Convert to PKCS8EncodedKeySpec
-            PKCS8EncodedKeySpec  keySpec = new PKCS8EncodedKeySpec(keyBytes);
-
-            // Step 3: Generate ECPrivateKey from KeyFactory
-            KeyFactory keyFactory = KeyFactory.getInstance("EC");
-            PrivateKey privateKey = keyFactory.generatePrivate(keySpec);
-
-            // Return as ECPrivateKey
-            encryptionPrivateKey = (ECPrivateKey) privateKey;
-        } catch (Exception e) {
-            LOGGER.error(e);
-        }
-
-
+        ECPrivateKey encryptionPrivateKey = myInfoSecurity.getEcPrivateKey(encryptionPrivateKeyString);
 
         LOGGER.info("Extracting payload");
-        String payload = myInfoSecurity.getPayload(encryptedResponse, encryptionPrivateKey);
+        String payload = getPayload(encryptedResponse, encryptionPrivateKey);
 
         LOGGER.info(payload);
 
         DecodedJWT personJWT = myInfoSecurity.decodeJwtToken(payload, jwksUrl);
-        LOGGER.info(new String (Base64.getDecoder().decode(personJWT.getHeader())));
-        LOGGER.info(new String (Base64.getDecoder().decode(personJWT.getPayload())));
 
         LOGGER.info("Formatting result");
-
         // Convert byte[] to String
         byte[] base64Decode = Base64.getDecoder().decode(personJWT.getPayload());
         String decryptedPersonApiResponse = new String(base64Decode);
 
         ObjectMapper objectMapper = new ObjectMapper();
+        return objectMapper.readValue(decryptedPersonApiResponse, DecryptedPersonInfo.class);
+    }
+
+    public String getPayload(String result, ECPrivateKey privateKey) {
+        JWEObject jweObject = null;
         try {
-            return objectMapper.readValue(decryptedPersonApiResponse, DecryptedPersonInfo.class);
-        } catch (JsonProcessingException e) {
+            // Parse JWE & validate headers
+            jweObject = EncryptedJWT.parse(result);
+
+            // Set PrivateKey and Decrypt
+            JWEDecrypter decrypter = new ECDHDecrypter(privateKey);
+            jweObject.decrypt(decrypter);
+
+        } catch (Exception e) {
             LOGGER.error(e);
-            return null;
         }
+        // Get String Payload
+
+        assert jweObject != null;
+
+        return jweObject.getPayload().toString();
     }
 }
